% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croneditor.R
\name{cron}
\alias{cron}
\title{Cron editor}
\usage{
cron(schedule, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{schedule}{Any valid cron schedule, for example "0 6 * * *"}

\item{width}{How wide should the editor be? Defaults to NULL, meaning 100\%}

\item{height}{How high should the editor be? Defaults to NULL, meaning 32px}

\item{elementId}{Optionally specifiy the ID of the element}
}
\description{
Cron editor
}
