% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croneditor.R
\name{croneditor}
\alias{croneditor}
\title{Cron Editor Input}
\usage{
croneditor(inputId, schedule = "* * * * *", label = NULL, language = "en-US")
}
\arguments{
\item{inputId}{The input ID for the cron editor.}

\item{schedule}{Initial schedule value (optional, default is "* * * * *").}

\item{label}{The label for the cron editor input.}

\item{language}{The language for the UI, can be one of en-US or nl-NL}
}
\value{
A UI element for Shiny.
}
\description{
This function creates a cron editor input widget in Shiny.
}
